Add AVI extension support and change logic in aria2c
When a new URI is added with magnet link, aria2c returns the GID of
the metalink. After the download is finished, the file(s) is(are)
downloaded in a new process, identified by the tag 'followed_by' on
the metalink process. Then we start watching the video file proccess.

When the download is finished, the status remain active (for seeding
purpose probably but it never seeds - may be a bug). Because of that,
The seed-time is set to 0 until I figure out another approach.